backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosmtplib>=3.0.1",
    "bcrypt>=4.0.1,<5.0.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "passlib>=1.7.4",
    "pillow>=10.4.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]
